refactor(viewer): extract <OrbitScene> mid-layer from OrbitViewer
Introduce a public <OrbitScene> for mounting the orbit scene graph inside a
caller-provided @react-three/fiber Canvas, sitting between the batteries-
included <OrbitViewer> and the internal OrbitSceneContents renderer.

- types: add OrbitSceneDataProps base; OrbitViewerProps and OrbitSceneProps
  both extend it. OrbitViewer gains a `canvas` escape hatch (camera/gl
  overrides); both gain `controls` (boolean | OrbitControls config) and `axes`.
- OrbitScene: owns the public-props -> renderer adaptation (frame resolution,
  trail buffers, arika WASM init, satellite maps) that previously lived in
  OrbitViewer, then renders OrbitSceneContents. No Canvas, no wrapper element.
- OrbitViewer: now a thin <div><Canvas><OrbitScene/></Canvas> wrapper.
- OrbitSceneContents: `controls`/`axes` gate the default camera package
  (OrbitControls + camera rig) and the axes helper; both default true, so the
  app's Scene path is unchanged.
- barrel: export OrbitScene, the new prop types, and SCENE_UP (needed to
  initialise a bring-your-own Canvas camera).
- regenerate the shadcn registry (orbit-viewer item: 38 -> 39 files).

Behaviour-preserving: OrbitViewer and the app render identically; no app
migration yet (that is the next step).
